Update sync interface.

+++ dingtalk/csharp/core/doc_1_0/Models/GetRangeResponseBody.cs
@@ -9,6 +9,11 @@ using Tea;
 namespace AlibabaCloud.SDK.Dingtalkdoc_1_0.Models
 {
     public class GetRangeResponseBody : TeaModel {
+        /// <summary>
+        /// 背景颜色
+        /// 最大size:
+        /// 	1000
+        /// </summary>
         [NameInMap("backgroundColors")]
         [Validation(Required=false)]
         public List<List<GetRangeResponseBodyBackgroundColors>> BackgroundColors { get; set; }
@@ -18,21 +23,21 @@ namespace AlibabaCloud.SDK.Dingtalkdoc_1_0.Models
             /// </summary>
             [NameInMap("red")]
             [Validation(Required=false)]
-            public long? Red { get; set; }
+            public int? Red { get; set; }
 
             /// <summary>
             /// RGB值中的绿色值
             /// </summary>
             [NameInMap("green")]
             [Validation(Required=false)]
-            public long? Green { get; set; }
+            public int? Green { get; set; }
 
             /// <summary>
             /// RGB值中的蓝色值
             /// </summary>
             [NameInMap("blue")]
             [Validation(Required=false)]
-            public long? Blue { get; set; }
+            public int? Blue { get; set; }
 
             /// <summary>
             /// 16进制表示的颜色
@@ -43,16 +48,28 @@ namespace AlibabaCloud.SDK.Dingtalkdoc_1_0.Models
 
         }
 
+        /// <summary>
+        /// 展示值
+        /// 最大size:
+        /// 	1000
+        /// </summary>
         [NameInMap("displayValues")]
         [Validation(Required=false)]
         public List<List<string>> DisplayValues { get; set; }
 
+        /// <summary>
+        /// 公式
+        /// 最大size:
+        /// 	1000
+        /// </summary>
         [NameInMap("formulas")]
         [Validation(Required=false)]
         public List<List<string>> Formulas { get; set; }
 
         /// <summary>
         /// 值
+        /// 最大size:
+        /// 	1000
         /// </summary>
         [NameInMap("values")]
         [Validation(Required=false)]
